Fix bug in documentation for default_goal_tolerance
Despite the documentation and debug messages, this value is assumed to be in radiance and is converted .toDeg (from radians) on line 476 of pre_computed_joint_trajectory_action_server.cpp. This commit updates documentation, debug messages, and default value accordingly. Another possible fix would be to change the code to match the documentation instead (and not convert this value from degrees to radians on line 476 of pre_computed_joint_trajectory_action_server.cpp). However I think this is the better fix, since the client is already using radians inside the trajectory_msgs/JointTrajectory msg anyway, so this reduces the number of units the client has to think in. (#62)